+++ src/drvizer/__init__.py
@@ -2,7 +2,7 @@
 Genomic Visualization Tools
 """
 
-__version__ = "1.0.0"
+__version__ = "0.1.0"
 
 from .api import DrViz
 from ._parallel import ParallelCoverageError